+++ kob-backend/src/main/java/com/sdu/kob/controller/PostController.java
@@ -23,7 +23,7 @@ public class PostController {
 
     @RequestMapping(value = "/api/post/getPosts/", method = RequestMethod.GET)
     public List<Post> getUserPosts(@RequestParam Map<String, String> data) {
-        Integer userId = Integer.parseInt(data.get("user_id"));
+        Long userId = Long.parseLong(data.get("user_id"));
         return postService.getAllPosts(userId);
     }
