+++ kob-backend/src/main/java/com/sdu/kob/service/MatchService.java
@@ -1,5 +1,5 @@
 package com.sdu.kob.service;
 
 public interface MatchService {
-    String startGame(Integer aId, Integer bId);
+    String startGame(Long aId, Long bId);
 }
