+++ kob-backend/src/main/java/com/sdu/kob/service/RecordService.java
@@ -3,7 +3,7 @@ package com.sdu.kob.service;
 import com.alibaba.fastjson.JSONObject;
 
 public interface RecordService {
-    JSONObject getAllRecords(Integer userId, Integer page);
+    JSONObject getAllRecords(Long userId, Integer page);
 
-    JSONObject getMyRecords(Integer userId, Integer page);
+    JSONObject getMyRecords(Long userId, Integer page);
 }
