+++ kob-backend/src/main/java/com/sdu/kob/service/StarPostService.java
@@ -1,7 +1,7 @@
 package com.sdu.kob.service;
 
 public interface StarPostService {
-    String starPost(Integer userId, Integer postId);
+    String starPost(Long userId, Long postId);
 
-    String unstarPost(Integer userId, Integer postId);
+    String unstarPost(Long userId, Long postId);
 }
