+++ kob-backend/src/main/java/com/sdu/kob/service/impl/FriendServiceImpl.java
@@ -28,8 +28,8 @@ public class FriendServiceImpl implements FriendService {
     public String follow(String friendName, String userName) {
         User user = userDAO.findByUserName(userName);
         User friend = userDAO.findByUserName(friendName);
-        Integer userId = user.getId();
-        Integer friendId = friend.getId();
+        Long userId = user.getId();
+        Long friendId = friend.getId();
         Friend relationship = friendDAO.findByUserAAndUserB(userId, friendId);
         if (relationship == null) {
             Friend relation = new Friend(userId, friendId, "true");
@@ -44,8 +44,8 @@ public class FriendServiceImpl implements FriendService {
     public String unfollow(String friendName, String userName) {
         User user = userDAO.findByUserName(userName);
         User friend = userDAO.findByUserName(friendName);
-        Integer userId = user.getId();
-        Integer friendId = friend.getId();
+        Long userId = user.getId();
+        Long friendId = friend.getId();
         friendDAO.update(userId, friendId, "false");
         return "success";
     }
@@ -53,14 +53,14 @@ public class FriendServiceImpl implements FriendService {
     @Override
     public JSONObject getUserFollowed(String userName) {
         User user = userDAO.findByUserName(userName);
-        Integer userId = user.getId();
+        Long userId = user.getId();
         List<Friend> followed = friendDAO.findByUserAAndFollowed(userId, "true");
         JSONObject resp = new JSONObject();
         List<JSONObject> items = new LinkedList<>();
         for (Friend friend : followed) {
             JSONObject item = new JSONObject();
-            Integer followedId = friend.getUserB();
-            User u = userDAO.findById((int) followedId);
+            Long followedId = friend.getUserB();
+            User u = userDAO.findById((long) followedId);
             item.put("id", u.getId());
             item.put("username", u.getUserName());
             item.put("state", checkLogin(u.getId()));
@@ -77,14 +77,14 @@ public class FriendServiceImpl implements FriendService {
     @Override
     public JSONObject getAllFollowers(String userName) {
         User user = userDAO.findByUserName(userName);
-        Integer userId = user.getId();
+        Long userId = user.getId();
         List<Friend> followers = friendDAO.findByUserBAndFollowed(userId, "true");
         JSONObject resp = new JSONObject();
         List<JSONObject> items = new LinkedList<>();
         for (Friend friend : followers) {
             JSONObject item = new JSONObject();
-            Integer followedId = friend.getUserA();
-            User u = userDAO.findById((int) followedId);
+            Long followedId = friend.getUserA();
+            User u = userDAO.findById((long) followedId);
             item.put("id", u.getId());
             item.put("username", u.getUserName());
             item.put("state", checkLogin(u.getId()));
@@ -101,18 +101,18 @@ public class FriendServiceImpl implements FriendService {
     @Override
     public JSONObject getFriends(String userName) {
         User user = userDAO.findByUserName(userName);
-        Integer userId = user.getId();
+        Long userId = user.getId();
         List<Friend> followers = friendDAO.findByUserBAndFollowed(userId, "true");
         List<Friend> followed = friendDAO.findByUserAAndFollowed(userId, "true");
 
         JSONObject resp = new JSONObject();
         List<JSONObject> items = new LinkedList<>();
         for (Friend a : followed) {
-            Integer userAId = a.getUserB();
+            Long userAId = a.getUserB();
             for (Friend b : followers) {
                 if (b.getUserA().equals(userAId)) {
                     JSONObject item = new JSONObject();
-                    User u = userDAO.findById((int) b.getUserA());
+                    User u = userDAO.findById((long) b.getUserA());
                     item.put("id", u.getId());
                     item.put("username", u.getUserName());
                     item.put("avatar", u.getAvatar());
@@ -130,7 +130,7 @@ public class FriendServiceImpl implements FriendService {
     }
 
     @Override
-    public JSONObject getRelationship(Integer searchId, Integer userId) {
+    public JSONObject getRelationship(Long searchId, Long userId) {
         JSONObject resp = new JSONObject();
         Friend relationshipA = friendDAO.findByUserAAndUserB(userId, searchId);
         Friend relationshipB = friendDAO.findByUserAAndUserB(searchId, userId);
@@ -163,7 +163,7 @@ public class FriendServiceImpl implements FriendService {
         return resp;
     }
 
-    public int checkLogin(Integer id) {
+    public int checkLogin(Long id) {
         if (goUsers.containsKey(id)) return 1;
         return 0;
     }
